Rename regconfig to Training
