+++ src/routes.js
@@ -5,7 +5,7 @@ import Full from './containers/Full/';
 import Dashboard from './views/Dashboard/';
 import Upload from './views/Upload/Upload';
 import JobStatus from './views/JobStatus/JobStatus';
-import RemainingTimeTraining from './views/RemainingTimeTraining';
+import Training from './views/Training/Training';
 
 export default (
   <Router history={hashHistory}>
@@ -14,7 +14,7 @@ export default (
       <Route path="dashboard" name="Dashboard" component={Dashboard}/>
       <Route path="upload" name="Upload" component={Upload}/>
       <Route path="jobStatus" name="Job status" component={JobStatus}/>
-      <Route path="regconfig" name="Remaining time training" component={RemainingTimeTraining}/>
+      <Route path="training" name="Training" component={Training}/>
     </Route>
   </Router>
 );
