+++ src/views/Training/Training.js
@@ -1,11 +1,11 @@
 import React, {Component} from 'react';
 import {connect} from 'react-redux';
 import PropTypes from 'prop-types';
-import {logListRequested} from '../actions/LogActions';
-import RemainingTimeCard from '../components/RemainingTimeCard';
-import {submitTraining} from '../actions/JobActions';
+import {logListRequested} from '../../actions/LogActions';
+import TrainingFormCard from '../../components/TrainingFormCard';
+import {submitTraining} from '../../actions/JobActions';
 
-class RemainingTimeTraining extends Component {
+class Training extends Component {
   componentDidMount() {
     // TODO refactor this
     if (this.props.logNames.length === 0) {
@@ -19,15 +19,15 @@ class RemainingTimeTraining extends Component {
     return (
       <div className="md-grid">
         <div className="md-cell md-cell--12">
-          <RemainingTimeCard logNames={this.props.logNames} fetchState={this.props.fetchState}
-                             onSubmit={this.props.onSubmitTraining}/>
+          <TrainingFormCard logNames={this.props.logNames} fetchState={this.props.fetchState}
+                            onSubmit={this.props.onSubmitTraining}/>
         </div>
       </div>
     );
   }
 }
 
-RemainingTimeTraining.propTypes = {
+Training.propTypes = {
   logNames: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
   onRequestLogList: PropTypes.func.isRequired,
   onSubmitTraining: PropTypes.func.isRequired,
@@ -47,4 +47,4 @@ const mapDispatchToProps = (dispatch) => ({
   onSubmitTraining: (endpoint, payload) => dispatch(submitTraining({endpoint, payload}))
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(RemainingTimeTraining);
+export default connect(mapStateToProps, mapDispatchToProps)(Training);
